make cart

+++ src/app/pages/cart/cart.page.ts
@@ -11,27 +11,43 @@ import { Product } from '../../models/product.model';
 })
 export class CartPage implements OnInit {
     public productCart: Array<Product> = [];
-
+  public total : number =0;
 
   constructor(public storage: CrudStorageService, public alertController: AlertController) { }
 
   async ngOnInit() {
-    this.productCart = await this.storage.read('product');
+    this.productCart = await this.storage.read('productcart');
+    this.total =  this.productCart.reduce(function(a, b) {
+        return a + b.soluongcart*b.price;
+      }, 0);
+    console.log(this.total)
   }
 
+
  
+  tangsoluong(productid){
+    let productupdate = this.productCart.find(element => element.id == productid)
+    productupdate.soluongcart++;
+    this.storage.update(productupdate)
+    this.total =  this.productCart.reduce(function(a, b) {
+        return a + b.soluongcart*b.price;
+      }, 0);
     
+  }
 
 
    deleteCartproduct(cartid) {
-    let deletefunction = () => {
+      let productupdate = this.productCart.find(element => element.id == cartid)
+    productupdate.soluongcart--;
+    this.storage.update(productupdate)
+    this.total =  this.productCart.reduce(function(a, b) {
+        return a + b.soluongcart*b.price;
+      }, 0);
+    if(productupdate.soluongcart < 1){
+      document.getElementById(productupdate.id).style.display= "none"
       this.storage.delete(cartid);
-    };
-    this.presentAlertMultipleButtons(
-      deletefunction,
-      "Xóa sản phẩm",
-      "Bạn có muốn xóa sản phẩm không."
-    );
+    
+    }
 
     // alert("da");
     // window.location.href = "/dashboard";
